Refactoring code
Put scripts into modules, fix minor errors
